refactor/deprecate_sql
remove the sql backend completely

If there is interest in alternative db backends they can be pluginified, for now it's easier to work across tools if we stick to json and XDG locations
